+++ host/class/cdc/usb_host_cdc_acm/test_app/pytest_usb_host_cdc.py
@@ -20,8 +20,8 @@ def test_usb_host_cdc(dut: Tuple[IdfDut, IdfDut]) -> None:
     tests = [
         # Device test mode    Host test case group
         ("dual_iface",        "cdc_acm"),
-        #("suspend_dconn",     "host_suspend_dconn"),
-        #("resume_dconn",      "host_resume_dconn"),
+        ("suspend_dconn",     "host_suspend_dconn"),
+        ("resume_dconn",      "host_resume_dconn"),
     ]
 
     for dev_test_mode, host_test_case_group in tests: